Add unit tests for private data APIs.
One of these tests has been failing for me, and I've traced the problem to a bug
in STAF 3.4.8. STAFStringCountSubStrings runs off the end of a buffer looking
for matches, and in the right circumstances will find a "match" in the trailing
junk after the legitimate data. This causes STAFEscapePrivacyDelimiters to
calculate the wrong size for its destination buffer, and leaves an extra byte at
the end of the final string.