Don't reference a MapClassDefinition in MapClass.
This addresses the problem of the map class and definition staying consistent by
basically copying the MapClassDefinition into the MapClass. You can create a
MapClassDefinition, use it to build a few MapClass instances, then change the
MapClassDefinition again without affecting the MapClass instances. Given a
MapClass instance, you can use .definition() to get back an equivalent
MapClassDefinition (though not identically the same as the original).